Recorder enhancements: Made capture interval fractional with adaptive scaling Improved layout of export dialog Added render time to final export screen

+++ plugins/dockers/recorder/recorder_config.cpp
@@ -46,12 +46,12 @@ void RecorderConfig::setSnapshotDirectory(const QString &value)
 }
 
 
-int RecorderConfig::captureInterval() const
+double RecorderConfig::captureInterval() const
 {
-    return config->readEntry(keyCaptureInterval, 1);
+    return config->readEntry(keyCaptureInterval, 1.);
 }
 
-void RecorderConfig::setCaptureInterval(int value)
+void RecorderConfig::setCaptureInterval(double value)
 {
     config->writeEntry(keyCaptureInterval, value);
 }
